Add BC6 support to nvtt lib and utils. - Use 3x3 eigensolver for initial fit in ZOH.  Slightly better perf and RMSE than power method. - Remove use of double precision in ZOH - speeds up by 12%. - Fixed RGBM encoding that was broken for HDR images. - Use gamma-2.0 space for RGBM for HDR images (improves precision in darks). - Use UNORM instead of TYPELESS formats when saving a DX10 .dds file.  The TYPELESS formats break most viewers. - Cleaned up warnings in ZOH code. - Command-line utils will warn if you give them an unrecognized parameter. - Added VS2010 profiling results.

+++ src/nvtt/bc6h/zoh.cpp
@@ -29,8 +29,8 @@ void ZOH::compress(const Tile &t, char *block)
 {
 	char oneblock[ZOH::BLOCKSIZE], twoblock[ZOH::BLOCKSIZE];
 
-	double mseone = ZOH::compressone(t, oneblock);
-	double msetwo = ZOH::compresstwo(t, twoblock);
+	float mseone = ZOH::compressone(t, oneblock);
+	float msetwo = ZOH::compresstwo(t, twoblock);
 
 	if (mseone <= msetwo)
 		memcpy(block, oneblock, ZOH::BLOCKSIZE);
